+++ BeeKeeper/src/com/beekeeper/controller/MainControllerServices.java
@@ -25,7 +25,7 @@ public interface MainControllerServices
 	public void notifyDeath(Agent a);
 
 	public void notifyAgentContact(int id1, int id2, double amount);
-	public HashMap<Integer, Integer> getAgentContacts();
+	public HashMap<Integer, Double> getAgentContacts();
 	public void freeLockAgentContacts();
 	
 	public void switchFrames(int index1, int index2);
